Major Connection Changes + Timeout Checking
Changed the DataType of Connections to a std::map, where a string of the IPs and Ports forms the key. Also added a thread into the CaptureEngine to manage the Connection Timeout checking function. Currently this build contains errors every once and awhile with the CreateOrUpdateConnections function. Added exception handler where it was breaking.