Graphical Bugs
Reordered when the outer circle is drawn, so that it appears behind the hosts. Changed when we mutex lock on the visualconnections to just before we replace the list, making the drawing more responsive. Fixed the Host Node size calculation, which now draws all hosts correctly in the circle without breaks.